restructured code and folder structure to get rid of import-caused errors, added favicon to get rid of 404 favicon not found error
