--- app/templates/base.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <body>
-        <h1>Categories</h1>
-        {% for c in categories %}
-            {{c.id}} | {{c.name}}
-            <br>
-        {% endfor %}
-    </body>
-</html>
